+++ frontend/src/components/ui/PrevProductItem.tsx
@@ -1,4 +1,3 @@
-import BuyerProductModel from '@/src/models/BuyerProductModel';
 import styled from '@emotion/styled';
 
 const Item = styled.div`
@@ -39,16 +38,29 @@ const Price = styled.p`
     font-weight: bold;
 `;
 
-export default function ProductItem({ product }: { product: BuyerProductModel }) {
+type ProductItemProps = {
+    product: {
+        id: string;
+        name: string;
+        grade: string;
+        amount: number;
+        price: number | null;
+        thumbnail: string;
+    };
+};
+
+export default function PrevProductItem({ product }: ProductItemProps) {
     return (
         <Item>
-            <Thumbnail src={product.images[0]} />
+            <Thumbnail src={product.thumbnail} />
             <ContentContainer>
                 <Title>{product.name}</Title>
                 <GradeAndAmount>
-                    {product.grade} 등급 ∙ {product.quantity} 개
+                    {product.grade} ∙ {product.amount}개
                 </GradeAndAmount>
-                <Price>{product.price} 원</Price>
+                <Price>
+                    {product.price !== null ? `${product.price.toLocaleString()}원` : '가격 미정'}
+                </Price>
             </ContentContainer>
         </Item>
     );
