+++ frontend/src/data/productDummyData.json
@@ -12,7 +12,8 @@
         "upload_date": "2025-03-09T14:30:00.000Z",
         "buyer": null,
         "purchase_date": null,
-        "sales_status": "available"
+        "sales_status": "available",
+        "origin_url": "https://naver.com"
     },
     {
         "id": 102,
@@ -29,7 +30,8 @@
         "upload_date": "2025-03-08T12:00:00.000Z",
         "buyer": null,
         "purchase_date": null,
-        "sales_status": "available"
+        "sales_status": "available",
+        "origin_url": "https://naver.com"
     },
     {
         "id": 103,
@@ -46,7 +48,8 @@
         "upload_date": "2025-03-07T10:45:00.000Z",
         "buyer": null,
         "purchase_date": null,
-        "sales_status": "available"
+        "sales_status": "available",
+        "origin_url": "https://naver.com"
     },
     {
         "id": 104,
@@ -63,7 +66,8 @@
         "upload_date": "2025-03-06T16:20:00.000Z",
         "buyer": null,
         "purchase_date": null,
-        "sales_status": "available"
+        "sales_status": "available",
+        "origin_url": "https://naver.com"
     },
     {
         "id": 105,
@@ -80,7 +84,8 @@
         "upload_date": "2025-03-05T14:00:00.000Z",
         "buyer": null,
         "purchase_date": null,
-        "sales_status": "available"
+        "sales_status": "available",
+        "origin_url": "https://naver.com"
     },
     {
         "id": 106,
@@ -97,7 +102,8 @@
         "upload_date": "2025-03-04T09:30:00.000Z",
         "buyer": null,
         "purchase_date": null,
-        "sales_status": "available"
+        "sales_status": "available",
+        "origin_url": "https://naver.com"
     },
     {
         "id": 107,
@@ -114,7 +120,8 @@
         "upload_date": "2025-03-03T08:20:00.000Z",
         "buyer": null,
         "purchase_date": null,
-        "sales_status": "available"
+        "sales_status": "available",
+        "origin_url": "https://naver.com"
     },
     {
         "id": 108,
@@ -129,7 +136,8 @@
         "upload_date": "2025-03-02T07:15:00.000Z",
         "buyer": null,
         "purchase_date": null,
-        "sales_status": "available"
+        "sales_status": "available",
+        "origin_url": "https://naver.com"
     },
     {
         "id": 109,
@@ -144,6 +152,7 @@
         "upload_date": "2025-03-01T12:00:00.000Z",
         "buyer": null,
         "purchase_date": null,
-        "sales_status": "available"
+        "sales_status": "available",
+        "origin_url": "https://naver.com"
     }
 ]
