+++ frontend/src/models/ProductModel.ts
@@ -1,77 +1,82 @@
 export default class ProductModel {
     id: number | null;
-    categoryId: number | null;
+    category: number | null;
     images: string[];
     name: string | null;
     description: string | null;
     grade: string | null;
     quantity: number;
     price: number;
-    sellerId: number | null;
-    uploadDate: string | null; // ISO 8601: YYYY-MM-DDTHH:MM:SS.sssZ
-    buyerId: number | null;
-    purchaseDate: string | null; // ISO 8601: YYYY-MM-DDTHH:MM:SS.sssZ
+    seller: number | null;
+    uploadDate: string | null;
+    buyer: number | null;
+    purchaseDate: string | null;
     salesStatus: string | null;
+    originUrl: string | null;
 
     constructor({
         id = null,
         images = [],
-        categoryId = null,
+        category = null,
         name = null,
         grade = null,
         quantity = 0,
         price = 0,
-        sellerId = null,
+        seller = null,
         uploadDate = null,
-        buyerId = null,
+        buyer = null,
         purchaseDate = null,
         description = null,
         salesStatus = null,
+        originUrl = null,
     }: {
         id?: number | null;
         images?: string[];
-        categoryId?: number | null;
+        category?: number | null;
         name?: string | null;
         grade?: string | null;
         quantity?: number;
         price?: number;
-        sellerId?: number | null;
+        seller?: number | null;
         uploadDate?: string | null;
-        buyerId?: number | null;
+        buyer?: number | null;
         purchaseDate?: string | null;
         description?: string | null;
         salesStatus?: string | null;
+        originUrl?: string | null;
     }) {
         this.id = id;
         this.images = images;
-        this.categoryId = categoryId;
+        this.category = category;
         this.name = name;
         this.grade = grade;
         this.quantity = quantity;
         this.price = price;
-        this.sellerId = sellerId;
+        this.seller = seller;
         this.uploadDate = uploadDate;
-        this.buyerId = buyerId;
+        this.buyer = buyer;
         this.purchaseDate = purchaseDate;
         this.description = description;
         this.salesStatus = salesStatus;
+        this.originUrl = originUrl;
     }
 
     static fromJson(jsonData: any): ProductModel {
         return new ProductModel({
-            id: jsonData.id,
-            images: jsonData.images || [],
-            categoryId: jsonData.categoryId,
-            name: jsonData.name,
-            grade: jsonData.grade,
-            quantity: jsonData.quantity,
-            price: jsonData.price,
-            sellerId: jsonData.sellerId,
-            uploadDate: jsonData.uploadDate,
-            buyerId: jsonData.buyerId,
-            purchaseDate: jsonData.purchaseDate,
-            description: jsonData.description,
-            salesStatus: jsonData.salesStatus,
+            id: jsonData['id'],
+            images: jsonData['images'] || [],
+            category: jsonData['category'],
+            name: jsonData['name'],
+            grade: jsonData['grade'],
+            quantity: jsonData['quantity'],
+            price: jsonData['price'],
+            seller: jsonData['seller'],
+            uploadDate: jsonData['upload_date'],
+            buyer: jsonData['buyer'],
+            purchaseDate: jsonData['purchase_date'],
+            description: jsonData['description'],
+            salesStatus: jsonData['sales_status'],
+            originUrl: jsonData['origin_url'],
         });
     }
 
@@ -79,17 +84,36 @@ export default class ProductModel {
         return {
             id: this.id,
             images: this.images,
-            categoryId: this.categoryId,
+            category: this.category,
             name: this.name,
             grade: this.grade,
             quantity: this.quantity,
             price: this.price,
-            sellerId: this.sellerId,
-            uploadDate: this.uploadDate,
-            buyerId: this.buyerId,
-            purchaseDate: this.purchaseDate,
+            seller: this.seller,
+            upload_date: this.uploadDate,
+            buyer: this.buyer,
+            purchase_date: this.purchaseDate,
             description: this.description,
-            salesStatus: this.salesStatus,
+            sales_status: this.salesStatus,
+            origin_url: this.originUrl,
+        };
+    }
+
+    toJsonWithoutId(): any {
+        return {
+            images: this.images,
+            category: this.category,
+            name: this.name,
+            grade: this.grade,
+            quantity: this.quantity,
+            price: this.price,
+            seller: this.seller,
+            upload_date: this.uploadDate,
+            buyer: this.buyer,
+            purchase_date: this.purchaseDate,
+            description: this.description,
+            sales_status: this.salesStatus,
+            origin_url: this.originUrl,
         };
     }
 }
