+++ frontend/src/services/productService.ts
@@ -1,41 +1,54 @@
 import axios from 'axios';
 import ProductModel from '../models/ProductModel';
 
-const API_BASE_URL = import.meta.env.VITE_BASE_URL;
-const IMAGE_BASE_URL = import.meta.env.VITE_IMAGE_BASE_URL;
+const API_BASE_URL = `${import.meta.env.VITE_BASE_URL}/products`;
 
 /**
- * 상품을 생성하고 서버에 저장합니다.
- * @param {ProductModel} product - 생성할 상품 객체
- * @returns {Promise<void>}
+ * 상품 리스트를 서버에서 요청하여 가져옵니다.
+ * 페이지네이션을 지원하며, nextPageUrl이 없으면 기본 URL에서 시작합니다.
+ * @param {string | null} nextPageUrl - 다음 페이지 URL (null이면 첫 페이지 요청)
+ * @param {number} pageSize - 페이지 크기
+ * @param {number} category - 카테고리
+ * @returns {Promise<{ results: ProductModel[]; next: string | null } | null>}
  */
-export const createProductInService = async (product: ProductModel): Promise<void> => {
+export const getProductListInService = async (
+    nextPageUrl: string | null,
+    pageSize: number,
+    category: number | null
+): Promise<{
+    results: ProductModel[];
+    next: string | null;
+} | null> => {
     try {
-        await axios.post(`${API_BASE_URL}`, product.toJson());
+        // api문서에서 필터가 categories로 되어있음
+        const params = { page_size: pageSize, categories: category };
+        const requestUrl = nextPageUrl ?? `${API_BASE_URL}/`;
+        const response = await axios.get(requestUrl, { params });
+        const data = response.data;
+        return {
+            results: data.results.map((product: any) => ProductModel.fromJson(product)),
+            next: data.next,
+        };
     } catch (error) {
-        console.error('Error creating product:', error);
-        throw error;
+        console.error('Error fetching product list:', error);
+        return null;
     }
 };
 
 /**
  * 상품을 생성하고 서버에 저장합니다.
- * @param {File} file - 생성할 상품 객체
- * @returns {Promise<string>}
+ * @param {ProductModel} product - 생성할 상품 객체
+ * @returns {Promise<ProductModel | null>}
  */
-export const uploadProductImageInService = async (file: File): Promise<string | null> => {
-    const formData = new FormData();
-    formData.append('image', file);
+export const createProductInService = async (
+    product: ProductModel,
+): Promise<ProductModel | null> => {
     try {
-        const response = await axios.post(`${IMAGE_BASE_URL}/upload/`, formData, {
-            headers: {
-                'Content-Type': 'multipart/form-data',
-            },
-        });
-        return response.data.image_url;
+        const response = await axios.post(`${API_BASE_URL}/`, product.toJsonWithoutId());
+        return ProductModel.fromJson(response.data);
     } catch (error) {
-        console.error('Error uploading product image:', error);
-        throw error;
+        console.error('Error creating product:', error);
+        return null;
     }
 };
 
@@ -44,9 +57,11 @@ export const uploadProductImageInService = async (file: File): Promise<string | 
  * @param {number} productId - 가져올 상품의 ID
  * @returns {Promise<ProductModel | null>}
  */
-export const getProductInService = async (productId: number): Promise<ProductModel | null> => {
+export const getProductInService = async (
+    productId: number,
+): Promise<ProductModel | null> => {
     try {
-        const response = await axios.get(`${API_BASE_URL}/${productId}`);
+        const response = await axios.get(`${API_BASE_URL}/${productId}/`);
         return ProductModel.fromJson(response.data);
     } catch (error) {
         console.error('Error fetching product:', error);
@@ -58,30 +73,32 @@ export const getProductInService = async (productId: number): Promise<ProductMod
  * 특정 productId로 서버에서 상품 데이터를 업데이트합니다.
  * @param {number} productId - 업데이트할 상품의 ID
  * @param {Partial<ProductModel>} updatedData - 업데이트할 데이터 객체
- * @returns {Promise<void>}
+ * @returns {Promise<ProductModel | null>}
  */
 export const updateProductInService = async (
     productId: number,
     updatedData: Partial<ProductModel>,
-): Promise<void> => {
+): Promise<ProductModel | null> => {
     try {
-        await axios.put(`${API_BASE_URL}/${productId}`, updatedData);
+        const response = await axios.put(`${API_BASE_URL}/${productId}/`, updatedData);
+        return ProductModel.fromJson(response.data);
     } catch (error) {
         console.error('Error updating product:', error);
-        throw error;
+        return null;
     }
 };
 
 /**
  * 특정 productId로 서버에서 상품 데이터를 삭제합니다.
  * @param {number} productId - 삭제할 상품의 ID
- * @returns {Promise<void>}
+ * @returns {Promise<boolean>}
  */
-export const deleteProductInService = async (productId: number): Promise<void> => {
+export const deleteProductInService = async (productId: number): Promise<boolean> => {
     try {
-        await axios.delete(`${API_BASE_URL}/${productId}`);
+        await axios.delete(`${API_BASE_URL}/${productId}/`);
+        return true;
     } catch (error) {
         console.error('Error deleting product:', error);
-        throw error;
+        return false;
     }
 };
